Move coprocessor functions to Arm and Arm64 modules, make them type-safe
