+++ crates/unicorn/src/lib.rs
@@ -648,82 +648,6 @@ impl<'a, D, A: UcArch> Unicorn<'a, D, A> {
             .and_then(|| Ok(value.into_boxed_slice()))
     }
 
-    /// Read ARM Coprocessor register
-    pub fn reg_read_arm_coproc(&self, reg: &mut RegisterARMCP) -> Result<(), uc_error> {
-        let curr_arch = self.get_arch();
-        match curr_arch {
-            #[cfg(feature = "arch_arm")]
-            Arch::ARM => {}
-            _ => return Err(uc_error::ARCH),
-        }
-
-        unsafe {
-            uc_reg_read(
-                self.get_handle(),
-                RegisterARM::CP_REG.into(),
-                core::ptr::from_mut(reg).cast(),
-            )
-        }
-        .into()
-    }
-
-    /// Write ARM Coprocessor register
-    pub fn reg_write_arm_coproc(&mut self, reg: &RegisterARMCP) -> Result<(), uc_error> {
-        let curr_arch = self.get_arch();
-        match curr_arch {
-            #[cfg(feature = "arch_arm")]
-            Arch::ARM => {}
-            _ => return Err(uc_error::ARCH),
-        }
-
-        unsafe {
-            uc_reg_write(
-                self.get_handle(),
-                RegisterARM::CP_REG.into(),
-                core::ptr::from_ref(reg).cast(),
-            )
-        }
-        .into()
-    }
-
-    /// Read ARM64 Coprocessor register
-    pub fn reg_read_arm64_coproc(&self, reg: &mut RegisterARM64CP) -> Result<(), uc_error> {
-        let curr_arch = self.get_arch();
-        match curr_arch {
-            #[cfg(feature = "arch_aarch64")]
-            Arch::ARM64 => {}
-            _ => return Err(uc_error::ARCH),
-        }
-
-        unsafe {
-            uc_reg_read(
-                self.get_handle(),
-                RegisterARM64::CP_REG.into(),
-                core::ptr::from_mut(reg).cast(),
-            )
-        }
-        .and(Ok(()))
-    }
-
-    /// Write ARM64 Coprocessor register
-    pub fn reg_write_arm64_coproc(&mut self, reg: &RegisterARM64CP) -> Result<(), uc_error> {
-        let curr_arch = self.get_arch();
-        match curr_arch {
-            #[cfg(feature = "arch_aarch64")]
-            Arch::ARM64 => {}
-            _ => return Err(uc_error::ARCH),
-        }
-
-        unsafe {
-            uc_reg_write(
-                self.get_handle(),
-                RegisterARM64::CP_REG.into(),
-                core::ptr::from_ref(reg).cast(),
-            )
-        }
-        .and(Ok(()))
-    }
-
     #[cfg(feature = "arch_arm")]
     fn value_size_arm64(curr_reg_id: i32) -> Result<usize, uc_error> {
         match curr_reg_id {
